// src/app/modules/document-deviation/update-bill-branch/models/update-bill-branch.model.ts
export interface BillBranchData {
    billId: number;
    billNumber: string;
    blngLevelId: number;
    blngLevelValue: string;
    status: string;
    customerName: string;
    submsnBrId: number;
    submsnBrName: string;
    colBrId: number;
    colBrName: string;
    altrBrId: number;
    altrBrName: string;
    updSubmsnBrId: number;
    updSubmsnBrName: string;
    updColBrId: number;
    updColBrName: string;
    updAltrBrId: number;
    updAltrBrName: string;
}